requirements.txt
streamlit 
feedparser 
transformers 
torch 
coqui-tts 
soundfile
